refactor(renderer): * a few more restructurings & tweaks * better defaults handling for rendering * etc
