+++ packages/utils/regl-renderer/src/index.js
@@ -2,9 +2,9 @@ module.exports = {
   prepareRender: require('./rendering/render'),
   drawCommands: {
     // draw commands should bootstrap themselves the first time they are run
-    drawGrid: require('./rendering/drawGrid/index.js'),
-    drawAxis: require('./rendering/drawAxis'),
-    drawMesh: require('./rendering/drawMesh/index.js')
+    drawGrid: require('./rendering/commands/drawGrid/index.js'),
+    drawAxis: require('./rendering/commands/drawAxis'),
+    drawMesh: require('./rendering/commands/drawMesh/index.js')
   },
   cameras: {
     camera: require('./cameras/camera'),
