+++ packages/utils/regl-renderer/src/rendering/commands/drawExps/drawText/index.js
