+++ packages/utils/regl-renderer/src/rendering/commands/drawMesh/colorOnlyShaders.js
